+++ src/HoneybeeDotNet/Model/Face3D.cs
@@ -42,7 +42,7 @@ namespace HoneybeeDotNet.Model
         /// <param name="type">type (default to &quot;Face3D&quot;).</param>
         /// <param name="holes">Optional list of lists with one list for each hole in the face.Each hole should be a list of at least 3 points and each point a list of 3 (x, y, z) values. If None, it will be assumed that there are no holes in the face..</param>
         /// <param name="plane">Optional Plane indicating the plane in which the face exists.If None, the plane will usually be derived from the boundary points..</param>
-        public Face3D(List<List<decimal>> boundary, string type = "Face3D", List<List<List<decimal>>> holes = default, Plane plane = default)
+        public Face3D(List<List<double>> boundary, string type = "Face3D", List<List<List<double>>> holes = default, Plane plane = default)
         {
             // to ensure "boundary" is required (not null)
             if (boundary == null)
@@ -73,7 +73,7 @@ namespace HoneybeeDotNet.Model
         /// <value>A list of points representing the outer boundary vertices of the face. The list should include at least 3 points and each point should be a list of 3 (x, y, z) values.</value>
         [DataMember(Name="boundary", EmitDefaultValue=false)]
         [JsonProperty("boundary")]
-        public List<List<decimal>> Boundary { get; set; }
+        public List<List<double>> Boundary { get; set; }
 
         /// <summary>
         /// Gets or Sets Type
@@ -88,7 +88,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Optional list of lists with one list for each hole in the face.Each hole should be a list of at least 3 points and each point a list of 3 (x, y, z) values. If None, it will be assumed that there are no holes in the face.</value>
         [DataMember(Name="holes", EmitDefaultValue=false)]
         [JsonProperty("holes")]
-        public List<List<List<decimal>>> Holes { get; set; }
+        public List<List<List<double>>> Holes { get; set; }
 
         /// <summary>
         /// Optional Plane indicating the plane in which the face exists.If None, the plane will usually be derived from the boundary points.
